+++ packages/base/src/utils.ts
@@ -31,6 +31,7 @@ export function findAssignedElement(slot: HTMLSlotElement, selector: string) {
   return null;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Constructor<T> = new (...args: any[]) => T;
 
 export function addHasRemoveClass(element: HTMLElement) {