+++ packages/ripple/src/ripple-directive.ts
@@ -54,6 +54,7 @@ let didApplyRippleStyle = false;
 const applyRippleStyle = () => {
   didApplyRippleStyle = true;
   const part = new NodePart({templateFactory});
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
   part.appendInto(document.head!);
   part.setValue(style);
   part.commit();
@@ -96,14 +97,14 @@ export const rippleNode = (options: RippleNodeOptions) => {
             interactionNode.removeEventListener(type, handler, applyPassive()),
     registerDocumentInteractionHandler:
         <K extends EventType>(evtType: K, handler: SpecificEventListener<K>) =>
-            document.documentElement!.addEventListener(
-                evtType, handler, applyPassive()),
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    document.documentElement!.addEventListener(
+        evtType, handler, applyPassive()),
     deregisterDocumentInteractionHandler: <K extends EventType>(
         evtType: string, handler: SpecificEventListener<K>) =>
-        document.documentElement!.removeEventListener(
-            evtType,
-            handler as EventListenerOrEventListenerObject,
-            applyPassive()),
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    document.documentElement!.removeEventListener(
+        evtType, handler as EventListenerOrEventListenerObject, applyPassive()),
     registerResizeHandler: (handler: SpecificEventListener<'resize'>) =>
         window.addEventListener('resize', handler),
     deregisterResizeHandler: (handler: SpecificEventListener<'resize'>) =>
@@ -129,13 +130,13 @@ export const ripple =
     directive((options: RippleOptions = {}) => (part: PropertyPart) => {
       const surfaceNode = part.committer.element as HTMLElement;
       const interactionNode = options.interactionNode || surfaceNode;
-      let rippleFoundation = (part.value as any);
+      let rippleFoundation = part.value as MDCRippleFoundation | typeof noChange;
       // if the interaction node changes, destroy and invalidate the foundation.
       const existingInteractionNode =
           rippleInteractionNodes.get(rippleFoundation);
       if (existingInteractionNode !== undefined &&
           existingInteractionNode !== interactionNode) {
-        rippleFoundation.destroy();
+        (rippleFoundation as MDCRippleFoundation).destroy();
         rippleFoundation = noChange;
       }
       // make the ripple, if needed
@@ -147,15 +148,15 @@ export const ripple =
         // otherwise update settings as needed.
       } else {
         if (options.unbounded !== undefined) {
-          rippleFoundation.setUnbounded(options.unbounded);
+          (rippleFoundation as MDCRippleFoundation).setUnbounded(options.unbounded);
         }
         if (options.disabled !== undefined) {
-          rippleFoundation.setUnbounded(options.disabled);
+          (rippleFoundation as MDCRippleFoundation).setUnbounded(options.disabled);
         }
       }
       if (options.active === true) {
-        rippleFoundation.activate();
+        (rippleFoundation as MDCRippleFoundation).activate();
       } else if (options.active === false) {
-        rippleFoundation.deactivate();
+        (rippleFoundation as MDCRippleFoundation).deactivate();
       }
     });