+++ packages/top-app-bar/src/mwc-top-app-bar-base.ts
@@ -80,7 +80,7 @@ export class TopAppBarBase extends BaseElement {
   // TODO(sorvell): MDC decorates the navigation icon and action items with
   // ripples. Since these are slotted items here, the assumption is that the
   // user brings a web component with a ripple if rippling is desired.
-  render() {
+  protected render() {
     return html`
       <header class="mdc-top-app-bar ${classMap(this.barClasses)}">
       <div class="mdc-top-app-bar__row">